+++ src/include/Graph_csr.h
@@ -11,35 +11,34 @@ class Graph_csr
 
 private:
     /// Number of vertices in the graph
-    int n;
+    uint32_t n;
 
     /// Number of edges in the graph
-    int m;
+    uint64_t m;
 
     /// Boolean value indicating whether the graph is directed or not
     bool directed;
 
-    int *o_First;      // index, in o_Target, where the outedges of vertex v starts. Size n
-    int *o_Target;     // second enpoint for each edge. Size m
-    int *o_degree;     // current out-degree for each vertex v. Size n
-    int *o_red_degree; // current red out-degree for each vertex v. Red degree is defined as the number of edges not propagated to all v's neighbours. Size n
+    uint64_t *o_First;      // index, in o_Target, where the outedges of vertex v starts. Size n
+    uint32_t *o_Target;     // second enpoint for each edge. Size m
+    uint32_t *o_degree;     // current out-degree for each vertex v. Size n
+    uint32_t *o_red_degree; // current red out-degree for each vertex v. Red degree is defined as the number of edges not propagated to all v's neighbours. Size n
 
-    int *i_First;      // index, in i_Target, where the inedges of vertex v starts.
-    int *i_Target;     // second enpoint for each edge
-    int *i_degree;     // current in-degree for each vertex v
-    int *i_red_degree; // current red in-degree for each vertex v. Red degree is defined as the number of edges not propagated to all v's neighbours
+    uint64_t *i_First;      // index, in i_Target, where the inedges of vertex v starts.
+    uint32_t *i_Target;     // second enpoint for each edge
+    uint32_t *i_degree;     // current in-degree for each vertex v
+    uint32_t *i_red_degree; // current red in-degree for each vertex v. Red degree is defined as the number of edges not propagated to all v's neighbours
 
     int k;
     float phi;
 
-    int *queue;
-    int *visited;
-    int bfs_timestamp;
+    uint32_t *queue;
+    uint32_t *visited;
+    uint32_t bfs_timestamp;
 
-    void process_edges(int *edges);
+    void process_edges(uint32_t *edges);
 
-    int propagate(int u);
-
+    int propagate(uint32_t u);
 
 public:
     T *balls;
@@ -57,7 +56,7 @@ public:
      * @param k: number of nodes to sample when updating the graph.
      * @param phi: threshold value for the red degree of a vertex.
      */
-    Graph_csr(int N, int M, bool isDirected = false, int k = 0, float phi = 0.0);
+    Graph_csr(uint32_t N, uint64_t M, bool isDirected = false, int k = 0, float phi = 0.0);
 
     /**
      * This is the constructor of the Graph_csr class for MinHashBall objects.
@@ -69,45 +68,45 @@ public:
      * @param n_hashes: number of hash functions to use for the MinHashBall objects.
      * @param hash_functions: array of hash functions to use for the MinHashBall objects.
      */
-    Graph_csr(int N, int M, bool isDirected, int k, float phi, int n_hashes, Hash<uint32_t> **hash_functions);
+    Graph_csr(uint32_t N, uint64_t M, bool isDirected, int k, float phi, int n_hashes, Hash<uint32_t> **hash_functions);
 
-    bool check_edge(int u, int v);
+    bool check_edge(uint32_t u, uint32_t v);
 
     /**
      * This method performs a breadth-first search on the graph starting from the given vertex u and stops at distance 2.
      * @param u: vertex to start the breadth-first search from.
      * @return: size of the subgraph reachable from vertex u within distance 2.
      */
-    int bfs_2(int u);
+    uint32_t bfs_2(uint32_t u);
 
     /**
      * This method returns the ball of radius 2 of the given vertex u.
      * @param u: vertex for which to return the ball of radius 2.
      * @return: vector containing the vertices in the ball of radius 2 of vertex u.
      */
-    std::vector<int> ball_2(int u);
+    std::vector<uint32_t> ball_2(uint32_t u);
 
     /**
      * This method returns the number of vertices in the graph.
      */
-    int getN() const;
+    uint32_t getN() const;
 
     /**
      * This method sets the number of vertices in the graph.
      * @param n: number of vertices in the graph.
      */
-    void setN(int n);
+    void setN(uint32_t n);
 
     /**
      * This method returns the number of edges in the graph.
      */
-    int getM() const;
+    uint64_t getM() const;
 
     /**
      * This method sets the number of edges in the graph.
      * @param m: number of edges in the graph.
      */
-    void setM(int m);
+    void setM(uint64_t m);
 
     /**
      * This method sets the threshold value for the red degree of each vertex in the graph.
@@ -125,7 +124,7 @@ public:
      * @param v: second endpoint of the edge to insert.
      * @return: void
      */
-    void insert_edge(int u, int v);
+    void insert_edge(uint32_t u, uint32_t v);
 
     /**
      * This method reads the graph from the given file and returns a Graph_csr object.
@@ -149,7 +148,7 @@ public:
      * @return: Graph_csr object representing the graph read from the edges.
      */
     template <typename U = T, typename std::enable_if<!std::is_same<U, MinHashBall>::value, int>::type = 0>
-    static Graph_csr<T> *from_edges(int *edges, int n, int m, bool isDirected, int k = 0, float phi = 0.0);
+    static Graph_csr<T> *from_edges(uint32_t *edges, uint32_t n, uint64_t m, bool isDirected, int k = 0, float phi = 0.0);
 
     /**
      * This method reads the graph from the given file and returns a Graph_csr<MinHashBall> object.
@@ -177,7 +176,7 @@ public:
      * @return: Graph_csr object representing the graph read from the edges.
      */
     template <typename U = T, typename std::enable_if<std::is_same<U, MinHashBall>::value, int>::type = 0>
-    static Graph_csr<T> *from_edges(int *edges, int n, int m, bool isDirected, int k = 0, float phi = 0.0, int n_hashes = 0, Hash<uint32_t> **hash_functions = NULL);
+    static Graph_csr<T> *from_edges(uint32_t *edges, uint32_t n, uint64_t m, bool isDirected, int k = 0, float phi = 0.0, int n_hashes = 0, Hash<uint32_t> **hash_functions = NULL);
 
     /**
      * This method computes minhash signatures for the ball of radius 2 of the given vertex u.
@@ -187,7 +186,7 @@ public:
      * @param n_hashes: number of hash functions to use.
      * @param hash_functions: array of hash functions to use.
      */
-    uint32_t *computeExactMHSignature(int u, int n_hashes, Hash<uint32_t> **hash_functions);
+    uint32_t *computeExactMHSignature(uint32_t u, int n_hashes, Hash<uint32_t> **hash_functions);
 
     /**
      * This method computes few-permutation hashing signatures for the ball of radius 2 of the given vertex u.
@@ -198,14 +197,14 @@ public:
      * @param hash_functions: array of hash functions to use.
      * @param sig_size: size of the signature to compute.
      */
-    uint32_t *computeExactOPHSignature(int u, int n_hashes, Hash<uint32_t> **hash_functions, int sig_size);
+    uint32_t *computeExactOPHSignature(uint32_t u, int n_hashes, Hash<uint32_t> **hash_functions, int sig_size);
 
     /**
      * This method updates the graph by adding the edge (u, v) to the graph.
      * @param u: first endpoint of the edge to add.
      * @param v: second endpoint of the edge to add.
      */
-    int update(int u, int v);
+    int update(uint32_t u, uint32_t v);
 
     /**
      * This method prints the graph to the standard output.
@@ -225,7 +224,7 @@ public:
      */
     void flush_graph();
 
-    void print_vertex(int i, bool doPrintBall = false);
+    void print_vertex(uint32_t i, bool doPrintBall = false);
 };
 
 #endif
