+++ src/include/Utils.h
@@ -6,10 +6,10 @@
 #include <unordered_map>
 #include <unordered_set>
 
-int* read_edges(std::string filename, int *n, int *m);
-void permute_edges(int* edges, int size);
+uint32_t *read_edges(std::string filename, uint32_t *n, uint64_t *m);
+void permute_edges(int *edges, int size);
 std::vector<std::pair<int, int>> read_ball_sizes(std::string filename);
-std::vector<std::pair<int, uint32_t*>> read_signatures(std::string filename, int sig_size);
+std::vector<std::pair<int, uint32_t *>> read_signatures(std::string filename, int sig_size);
 std::unordered_map<int, std::unordered_set<int>> read_balls(std::string filename);
 
-#endif 
+#endif
